requirements.txt
# Home Assistant Data Collection
# Core dependencies

# API & HTTP
requests>=2.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Optional: for export formats
# openpyxl>=3.1.0  # Excel support
# pyarrow>=12.0.0  # Parquet support
